Create custom FilterParams class and use it in TriageApiClient
jsonapi_client provides filtering functionality through its Filter
class, but that class takes conditions as kwargs, which only allows
for attributes and values.

We also want to support arbitrary operations, so we'll replace the
kwargs with a list of dicts having three keys: attribute, value, op.

Supported operations vary depending on the attribute, and can be found
in the Triage API documentation.