Add schema.json and pass it to jsonapi_client.Session()
jsonapi_client always performs validation for write requests, so it
needs a schema. We could try to bypass this requirement and just let
Triage do the validation, but for now this is easier.

We'll decide later if/how/where we want to maintain this schema.json
file. Maybe Triage itself could provide it.